+++ app/(gcforms)/[locale]/(form administration)/forms/page.tsx
@@ -180,6 +180,10 @@ export default async function Page(props: {
       overdue: false,
       ttl: ttl ? new Date(ttl) : null,
       hasEditLock: templateIdsWithEditLocks.has(id),
+      url: `/${locale}/id/${id}`,
+      isShared: userCount + pendingUserCount >= 2,
+      hasDraft: !!template.currentDraftVersionId,
+      versionNumber: template.versionNumber ?? null,
       collaboratorCount: {
         userCount,
         pendingUserCount,